:recycle: Rename hwWallet.js to hwManager.js
