+++ src/store/middlewares/hwManager.js
